Token auth

+++ Projekcik.Client/src/Services/AuthService.js
@@ -1,17 +1,17 @@
 import APIService from './APIService';
 
-
 export default class AuthService {
+  static register(userData) {
+    return APIService.post('api/Auth/register', userData);
+  }
 
-    static register (userData) {
-        return APIService.post('api/Auth/register', userData);
-    }
+  static login(userData) {
+    return APIService.post('api/Auth/authenticate', userData);
+  }
 
-    static login (userData) {
-        return APIService.post('api/Auth/authenticate', userData);
+  static logout() {
+    if (APIService.isAuthenticated()) {
+      window.localStorage.clear();
     }
-
-    static logout () {
-        window.localStorage.clear();
-    }
-};
+  }
+}
